# resources/skg-if-openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0-alpha
  title: SKG-IF OpenAPI
  termsOfService: 'https://skg-if.github.io/interoperability-framework/'
  contact:
    url: 'https://github.com/skg-if'
  license:
    name: CC-BY
    url: 'http://creativecommons.org/licenses/by/4.0/'
  x-logo:
    url: 'https://www.rd-alliance.org/wp-content/uploads/2024/04/RDA_Logotype_CMYK.png'
  description: |
    OpenAPI end points specifications for [SKG-IF](https://skg-if.github.io/interoperability-framework/)

    Last version of the SKG-IF OpenAPI is available at : https://skg-if.github.io/api/

    # Note for implementers :
      - Please refer to [SKG-IF OpenAPI Implementer documentation](https://docs.google.com/document/d/1t7b7h28UTtM56Sda4NGJIp0hnQfGbcVVGn12fny9wfI/edit?usp=sharing)
      - [RDA SKG-IF WG](https://skg-if.github.io/) does not provide any official implementation of the current specification.
      - Recommended OpenAPI viewer :
        - StopLight/Elements Viewer [Current/last SKG-IF OpenAPI version](https://elements-demo.stoplight.io/?spec=https://w3id.org/skg-if/api/skg-if-openapi.yaml).
        - Swagger : [Current/last SKG-IF OpenAPI version](https://editor-next.swagger.io/)
...
servers:
  - url: '{protocol}://{server}:{port}/{skg_if_api_path}'
    variables:
      protocol:
        enum:
          - "https"
          - "http"
        default: "https"
        description: service provider protocol.
      server:
        default: example.com
...
      port:
        default: "8443"
...
paths:
  '/products/{short_local_identifier}':
    get:
      tags:
          - Product
      summary: Get product by id
      description: |
          Get single `product`. See definition in SKG-IF [Research product](https://skg-if.github.io/interoperability-framework/docs/research-product.html) ( entity_type:product ).
      operationId: getProductById
      parameters:
        - $ref : '#/components/parameters/shortLocalIdPathParam'
      responses:
        '200':
          description: |
            Success.
          content:
            # not 'application/json-ld' to be compatible with StopLight PRISM tool
            application/json:
              schema:
                required: ['@context','@graph']
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}]
                  "@graph":
                    type: array
                    minItems: 1
                    maxItems: 1
                    items:
                      $ref: "#/components/schemas/Product"
              examples:
                JournalArticleEx01:
                  summary: "Journal article - FAIR Guiding Principles"
                  value :
                    "@context": [
                      "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                      "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                      {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}
                    ]
                    "@graph" :
                      - local_identifier: http://example.com/skg-if/api/products/prd-c66c6-38be-4d5f-85db-d44c9f869333
                        entity_type: product
                        product_type: literature
                        titles :
                          "en": ["The FAIR Guiding Principles for scientific data management and stewardship"]
                        abstracts :
                          "en": ["There is an urgent need to improve the infrastructure supporting the reuse of scholarly data. A diverse set of stakeholders—representing academia, industry, funding agencies, and scholarly publishers—have come together to design and jointly endorse a concise and measureable set of principles that we refer to as the FAIR Data Principles"]
                        identifiers:
                          - value: "10.1038/sdata.2016.18"
                            scheme: "doi"
                          - value: "26978244"
                            scheme: "pmid"
                        contributions:
                          - contribution_types:
                              - "writing – original draft"
                              - "conceptualization"
                            role: "author"
                            rank: 1
                            by:
                              name: "Mark Wilkinson"
                              given_name: "Mark"
                              family_name: "Wilkinson"
                              identifiers:
                              - value: "0000-0001-6960-357X"
                                scheme: "orcid"
                              local_identifier: http://example.com/skg-if/api/persons/otf___1730027051396___person-1
                              entity_type: "person"
                            declared_affiliations:
                              - name: "Center for Plant Biotechnology and Genomics, Universidad Politécnica de Madrid, Madrid, 28223, Spain"
                                local_identifier: http://example.com/skg-if/api/otf___1730027051396___person-1-aff-1
                                entity_type: "organisation"
                                country: "ES"
                        manifestations:
                          - type:
                              labels:
                                en: localtypey
                              defined_in: http://www.example.com/types/onto
                              class: http://www.example.com/types/onto/localtypemethodology
                            biblio:
                              hosting_data_source:
                                local_identifier: http://example.com/skg-if/api/arch-6f368a3a-b1cf-498f-b2de-27135d1e0075
                                entity_type: datasource
                                name: "An Archive"
                              in:
                                name: "Scientific Data"
                                acronym: "Sci Data"
                                identifiers:
                                  - "value": "2052-4463"
                                    "scheme": "issn"
                                local_identifier: http://example.com/skg-if/api/venues/journal-46bc1-00f8-42dd-8221-ac4deb52fb25
                                entity_type: "venue"
                            dates: # http://api.crossref.org/works/10.1038/sdata.2016.18
                              publication:
                                - "2016-03-15"
                              collected:
                                - "2013"
                                - "2014"
                              deposit:
                                - "2023-01-04"
                        related_products:
                          cites:
                            - "614a6575-5d6c-416c-a8e1-f49a5d589bf8"

                DatasetEx02:
                  summary: "Dataset - OpenAIRE Graph dataset new collected projects"
                  value :
                    "@context": [
                       "https://w3id.org/skg-if/context/1.1.0/skg-if.json",
                       "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json",
                       {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}
                    ]
                    "@graph" :
                      - local_identifier: http://example.com/skg-if/api/products/prd-c66c6-38be-4d5f-85db-d44c9f869555
                        entity_type: product
                        product_type: research data
                        titles :
                          "en": ["OpenAIRE Graph dataset - new collected projects"]
                        abstracts :
                          "en": ["The dataset includes metadata about projects grants collected by OpenAIRE since October 2024."]
                        identifiers:
                          - value: "10.5281/zenodo.14622592"
                            scheme: "doi"
                          - value: "W4393781519"
                            scheme: "openalex"
        '404':
          description: |
            Error if entity does not exists
...
components:
  securitySchemes:
    main_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://example.com/api/oauth/dialog'
          scopes:
            'read:users': read users info
            'write:users': modify or remove users
    api_key:
      type: apiKey
      in: header
      name: api_key
    basic_auth:
      type: http
      scheme: basic
  parameters:
    shortLocalIdPathParam:
      name: short_local_identifier
      in: path
      description: |
        entity id. URL suffix used in OpenAPI get by id operations. \
        `https://[your-server-root-skg-if-url]/[products|persons|organisations|datasources|venues|topics]/{short_local_identifier}` \
        example `http://example.com/skg-if/api/products/c66c6-38be-4d5f-85db-d44c9`
      required: true
      schema:
        type: string
#    filterQueryParam: # filterQueryParam duplicated in each search endpoint ( only way to add distinct examples for each endpoint )
#      name: filter
#      in: query
#      description: |
#            Search filter. Format : Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
#      schema:
#        type: string
#        pattern: '^(,?.+:.+)*$'
...
  schemas:
...
    JsonLdCtxDataModel:
      type: string
      description: "URL to SKG-IF data model context supported by the current OpenAPI `https://w3id.org/skg-if/context/1.1.0/skg-if.json`"
      enum:
        - https://w3id.org/skg-if/context/1.1.0/skg-if.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.1.0/skg-if.json']
    JsonLdCtxSearch:
      type: string
      description: "URL to SKG-IF API context (meta section) supported by the current OpenAPI `https://w3id.org/skg-if/context/1.0.0/skg-if-api.json`"
      enum:
        - https://w3id.org/skg-if/context/1.0.0/skg-if-api.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.0.0/skg-if-api.json']
    JsonLdCtxBaseOrMore:
      description : 'Additional context mappings ex: `{ "@base":"https://w3id.org/skg-if/sandbox/my-skg-acronym" }`'
      properties:
        "@base":
          type: string
          description : |
            Default base URL specific to the implementer.\
            common format : `https://w3id.org/skg-if/sandbox/{my-skg-acronym}` \
            example : `https://w3id.org/skg-if/sandbox/openaire/`
          x-faker:
            fake: ['https://w3id.org/skg-if/sandbox/{{random.alpha(5)}}/']
          examples:
            - "https://w3id.org/skg-if/sandbox/openaire/"
            - "https://w3id.org/skg-if/sandbox/cessda/"
            - "https://w3id.org/skg-if/sandbox/foobar/"
      additionalProperties:
        type: string
        description: "context URL of extension entities"
        x-faker:
          fake : ['http://www.example.com/skg-if/{{random.alpha(3)}}']